rename files test_*
