+++ aeon/distances/tests/test_distances.py
@@ -4,8 +4,8 @@ from numpy.testing import assert_almost_equal
 
 from aeon.distances import distance as compute_distance
 from aeon.distances._distance import DISTANCES
-from aeon.distances.tests._expected_results import _expected_distance_results
-from aeon.distances.tests._utils import create_test_distance_numpy
+from aeon.distances.tests.test_expected_results import _expected_distance_results
+from aeon.distances.tests.test_utils import create_test_distance_numpy
 
 
 def _validate_distance_result(x, y, name, distance, expected_result=10):
