+++ aeon/regression/tests/test_all_regressors.py
@@ -7,7 +7,7 @@ import numpy as np
 from sklearn.utils._testing import set_random_state
 
 from aeon.datasets import load_cardano_sentiment, load_covid_3month
-from aeon.regression.tests._expected_outputs import (
+from aeon.regression.tests.test_expected_outputs import (
     cardano_sentiment_preds,
     covid_3month_preds,
 )
