+++ aeon/transformations/tests/test_all_transformers.py
@@ -9,7 +9,7 @@ from sklearn.utils._testing import set_random_state
 
 from aeon.datasets import load_basic_motions, load_unit_test
 from aeon.tests.test_all_estimators import BaseFixtureGenerator, QuickTester
-from aeon.transformations.tests._expected_outputs import (
+from aeon.transformations.tests.test_expected_outputs import (
     basic_motions_result,
     unit_test_result,
 )
